+++ packages/terra-clinical-item-display/CHANGELOG.md
@@ -5,6 +5,10 @@
 * Added
   * `strong` tag around strong style text.
   * `span` tag around non-strong style text.
+  * Added `textStyleMeaning` prop to allow defining a meaning for styled text to be read by screen readers.
+ 
+* Deprecated
+  * Deprecated `isDisabled` prop as the feature does not meet a11y standards.
 
 ## 4.6.0 - (March 29, 2023)